Fixed deleting classes which did not completely happen. Added a Test for this. Fixed template field in inline.vm and edit.vm which was not taking the template field from the request and this lead to an exception on certain edits Fixed backwards compatibility issues with velocity variables (from skins) and with reduced URLs (options now need to be activated to turn on reduced URLs)
git-svn-id: https://svn.xwiki.org/svnroot/xwiki/xwiki/trunk@916 f329d543-caf0-0310-9063-dda96c69346f

+++ src/main/java/com/xpn/xwiki/store/XWikiHibernateStore.java
@@ -1081,7 +1081,7 @@ public class XWikiHibernateStore extends XWikiDefaultStore {
             }
 
             BaseClass bclass = doc.getxWikiClass();
-            if ((bclass==null)&&(bclass.getName()!=null)) {
+            if ((bclass.getFieldList().size()>0)&&(useClassesTable(true, context))) {
                 deleteXWikiClass(bclass, context, false);
             }
 
@@ -1370,11 +1370,12 @@ public class XWikiHibernateStore extends XWikiDefaultStore {
 
             if (!object.getClassName().equals("internal")) {
                 for (Iterator it = object.getFieldList().iterator(); it.hasNext();) {
-                    BaseProperty property = (BaseProperty)it.next();
+                    BaseElement property = (BaseElement) it.next();
                     if (!handledProps.contains(property.getName())) {
                         if (evict)
                             session.evict(property);
-                        session.delete(property);
+                        if (session.get(property.getClass(), property)!=null)
+                            session.delete(property);
                     }
                 }
             }
